requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.4
django-appconf==1.0.2
django-celery==3.2.2
django-extensions==1.9.9
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
django-rest-swagger==2.1.2
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djoser==1.1.5
djrill==2.1.0
factory-boy==2.10.0
fake-factory==0.7.2
Faker==0.8.10
idna==2.6
ipaddress==1.0.19
itypes==1.1.0
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.0
oauthlib==2.0.6
openapi-codec==1.3.2
Pillow==5.0.0
psycopg2==2.7.3.2
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2018.3
redis==2.10.6
requests==2.18.4
simplejson==3.13.2
six==1.11.0
sorl-thumbnail==12.4.1
text-unidecode==1.1
tqdm==4.19.5
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
